GCW-3451 Move test utils functions to Login
Functions are too Login component specific to deserve own folder
